--- _config.yml
@@ -1,1 +1,0 @@
-include: [".htaccess", "node_modules"]
