+++ assets/js/firebase.js
@@ -1,7 +1,3 @@
-// const $ = require("jquery");
-// import { insertData } from "../../index.js";
-// import axios from "axios";
-
 var firebaseConfig = {
   apiKey: "AIzaSyBjS1MGlsOhRi0LHvUCLtoP8yzv46G-YPg",
   authDomain: "brewcrawler-ae0b1.firebaseapp.com",
@@ -18,16 +14,16 @@ let returnedCrawlData = [];
 var crawlCode = 0;
 var passingCrawl = [];
 
-$("#crawl-submit").click(function () {
-  crawlCode = Math.round(Math.random() * 1000000);
-  passingCrawl = JSON.parse(localStorage.getItem("crawlArray"));
-  console.log(passingCrawl);
-  console.log("crawlCode" + crawlCode);
-  console.log("success?");
-  writeCrawl();
-  localStorage.setItem("crawlCode", JSON.stringify(crawlCode));
-  window.location.pathname = "./confirmation.html";
-});
+// $("#crawl-submit").click(function () {
+//   crawlCode = Math.round(Math.random() * 1000000);
+//   passingCrawl = JSON.parse(localStorage.getItem("crawlArray"));
+//   console.log(passingCrawl);
+//   console.log("crawlCode" + crawlCode);
+//   console.log("success?");
+//   writeCrawl();
+//   localStorage.setItem("crawlCode", JSON.stringify(crawlCode));
+//   window.location.pathname = "./confirmation.html";
+// });
 
 function writeCrawl() {
   firebase.database().ref(`crawl-code/${crawlCode}`).set({
@@ -58,19 +54,22 @@ const getById = async (id) => {
       returnedCrawlData.push(res.data);
     }
     console.log(JSON.parse(window.localStorage.crawlArray).length);
-    insertData(returnedCrawlData);
-    // console.log(returnedCrawl);
+    console.log(insertData);
+    //   insertData(returnedCrawlData);
   } catch (error) {
     console.error(error);
   }
 };
 
-if (window.location.pathname === "/confirmation.html") {
-  let crawlCode = JSON.parse(window.localStorage.getItem("crawlCode"));
-  let crawlData = JSON.parse(window.localStorage.getItem("crawlArray"));
-  for (var i = 0; i < crawlData.length; i++) {
-    getById(crawlData[i]);
+window.onload = () => {
+  if (window.location.pathname === "/confirmation.html") {
+    let crawlCode = JSON.parse(window.localStorage.getItem("crawlCode"));
+    let crawlData = JSON.parse(window.localStorage.getItem("crawlArray"));
+    for (var i = 0; i < crawlData.length; i++) {
+      getById(crawlData[i]);
+    }
+    console.log("returned array", returnedCrawlData);
+    document.getElementById("confirmation-code").innerHTML = crawlCode;
   }
-  console.log("returned array", returnedCrawlData);
-  document.getElementById("confirmation-code").innerHTML = crawlCode;
-}
+  console.log(insertData);
+};
