+++ confirmation.html
@@ -77,6 +77,7 @@
   <script src="https://www.gstatic.com/firebasejs/7.14.0/firebase-database.js"></script>
   <script src="https://cdnjs.cloudflare.com/ajax/libs/materialize/1.0.0/js/materialize.min.js"></script>
   <script>M.AutoInit()</script>
+  <script src="./sharedFunctions.js"></script>
   <script src="./index.js"></script>
   <script src="./assets/js/firebase.js"></script>
 </body>
